==> notebooks/1_Get_Data.ipynb <==
# Get Data

This notebook fetches the latest data for Gradio fine-tuning from various sources.

import os

this_dir = os.getcwd()

import shutil

import requests
from urllib.parse import urljoin

headers = None

# headers = {
#     "Authorization": f"Bearer {'your_token_here'}"
# }
# i got rate limited so had to go get a auth token at https://github.com/settings/tokens but you should be ok to run once without this

# get latest gradio repo python code
for this I am utilising this [github2file](https://github.com/cognitivecomputations/github2file) repo from the inspirational ehartford 

os.chdir(this_dir+"/../../github2file/") # assumed you cloned that repo next to this one

!python github2file.py https://github.com/gradio-app/gradio

shutil.copy("gradio_python.txt", this_dir + "/../data/latest-repo/gradio_repo_one_file.txt")

# get latest gradio repo docs / guides

os.chdir(this_dir+"/../data/latest-docs/")

# Define the base URL for the repository and the sub-directory
base_url = 'https://github.com/gradio-app/gradio/'
sub_directory = 'main/guides/'

# GitHub API base URL for fetching repository contents
api_base_url = 'https://api.github.com/repos/gradio-app/gradio/contents/guides'

# Directory to save downloaded files
save_directory = './'

# Create the save directory if it doesn't exist
if not os.path.exists(save_directory):
    os.makedirs(save_directory)

def download_file(file_url, save_path):
    response = requests.get(file_url)
    if response.status_code == 200:
        with open(save_path, 'wb') as file:
            file.write(response.content)
        print(f"Downloaded: {save_path}")
    else:
        print(f"Failed to download {file_url}")

def get_files_from_github(api_url,headers):
    if headers = None:
        response = requests.get(api_url)
    else:
        response = requests.get(api_url,headers=headers)
    if response.status_code == 200:
        return response.json()
    else:
        print(f"Failed to fetch repository contents: {response.status_code}")
        return []

def process_directory(api_url, local_dir,headers=None):
    contents = get_files_from_github(api_url,headers)
    
    for item in contents:
        if (item['type'] == 'dir') & ("cn" not in item['name']) & ("assets" not in item['name']): #ignore the chinese version - gonna finetune in english
            # Create a corresponding local directory
            new_local_dir = os.path.join(local_dir, item['name'])
            if not os.path.exists(new_local_dir):
                os.makedirs(new_local_dir)
            # Recursively process the sub-directory
            process_directory(item['url'], new_local_dir,headers)
        elif item['type'] == 'file' and item['name'].endswith('.md'):
            # Download the file
            download_file(item['download_url'], os.path.join(local_dir, item['name']))


process_directory(api_base_url, save_directory,headers)

# Chat and Existing User Queries
### TODO

import pandas as pd

from IPython.display import display, Markdown

import re

response = requests.get("https://raw.githubusercontent.com/gradio-app/awesome-demos/main/README.md")

readme_content = response.text

# Convert markdown to HTML
html_content = markdown(readme_content)

# Function to convert markdown table to pandas DataFrame
def markdown_table_to_dataframe(table):
    rows = table.strip().split('\n')
    header = rows[0]
    columns = [col.strip() for col in header.split('|') if col.strip()]
    data = []
    for row in rows[2:]:  # Skip the header and separator rows
        values = [val.strip() for val in row.split('|') if val.strip()]
        if values:
            data.append(values)
    return pd.DataFrame(data, columns=columns)

all_links = {}
for section in readme_content.split('##')[1:]:
        
    
    title, table = section.split('|',1)
    title=title.strip()
    table="|"+table
    
    
    all_links[title] = markdown_table_to_dataframe(table)

all_demo_links = pd.concat(all_links).reset_index()

link_regex=re.compile(r'\[.*?\]\((.*?)\)')

def get_link(link_row):
    raw_link = (link_regex.search(link_row['Demo name (link to demo)'])[1]+"/raw/main/app.py")
    
    try:
        return requests.get(raw_link).text
    except:
        print(raw_link)
        return 'FAILED ' + raw_link 
        

all_demo_links['python_str'] = all_demo_links.apply(lambda x: get_link(x),axis=1)

#display(Markdown(f"```python\n{all_demo_links.iloc[10]['python_str']}\n```"))

## Capture static status

link_row = all_demo_links.iloc[0]

def get_status(link_row):
    response = requests.get(link_regex.search(link_row['status badge'])[1])
    
    if "demo status: up" in str(response.content):
        return 'up'
    else:
        return 'drama'

all_demo_links['static_status'] = all_demo_links.apply(lambda x: get_status(x),axis=1)

all_demo_links.to_parquet( this_dir + "/../data/latest-demos/all_demo_python.parquet")

display(Markdown(f"```python\n{all_demo_links.iloc[62]['python_str']}\n```"))
