==> notebooks/5_predibase.ipynb <==
#!pip install predibase

from predibase import Predibase, FinetuningConfig, DeploymentConfig

pb = Predibase(api_token="asdf")

import os

dataset = pb.datasets.from_file(os.getcwd() + f'/../datasets/finetune_data_mistral_predibase.jsonl', name="gradio_test2")

# Create an adapter repository
repo = pb.repos.create(name="gradio-assistant", description="coding assistant for using gradio", exists_ok=True)

# Start a fine-tuning job, blocks until training is finished
adapter = pb.adapters.create(
    config=FinetuningConfig(
        base_model="mistral-7b"
    ),
    dataset=dataset, # Also accepts the dataset name as a string
    repo=repo,
    description="initial model with defaults"
)

adapter = pb.adapters.get("gradio-assistant/1")
adapter

input_prompt="<s>[INST] Write me a python script for a gradio app that can dynamically plot a scatter plot given a multi selection options ['a','b','c'] which is used to select a column of an input dataframe [/INST] "

lorax_client = pb.deployments.client("mistral-7b")

print(lorax_client.generate(input_prompt, adapter_id="gradio-assistant/1", max_new_tokens=1000).generated_text)
